Move login to component
Moves the main login page to a separate login component
